+++ pkg/services/authz/wireset.go
@@ -1,4 +1,4 @@
-package authorization
+package authz
 
 import (
 	"github.com/google/wire"
